+++ src/main/java/frc/states/SuperStructureStates/HighGoalState.java
@@ -8,7 +8,7 @@ public class HighGoalState extends SuperStructureState {
     }
 
     @Override
-    public double wristPosition() {
+    public int wristPosition() {
         return 4000;
     }
