+++ src/main/java/frc/states/SuperStructureStates/LowGoalState.java
@@ -8,7 +8,7 @@ public class LowGoalState extends SuperStructureState {
     }
 
     @Override
-    public double wristPosition() {
+    public int wristPosition() {
         return 5900;
     }
