+++ src/main/java/frc/states/SuperStructureStates/SuperStructureState.java
@@ -2,5 +2,5 @@ package frc.states.SuperStructureStates;
 import fireLib.stateControl.*;
 public abstract class SuperStructureState extends State {
     public abstract int armPosition();
-    public abstract double wristPosition();
+    public abstract int wristPosition();
 }
